On inject return string and JSON DOM of selection
On injection, the string representation of the user selected content, as
well as the JSON string representation of the first element in the
selection are returned in a object. (1)

After injection (2), the returned object is inspected for its properties,
if any of them should be empty, a message wil be shown to the user
telling them nothing has been selected. (3)

The JSON string representation of the anchorNode is parsed, and set as
the inner contents of richText div. Some variable names have changed.

1. content_scripts/get_selection.js
2. popup.js
3. popup.html