Selected DOMstring sanitized before storage
The selection is no longer saved at content script level (1), instead it
is sanitized using DOMpurify on popup.js execution (2).

DOMpurify script tag was added to popup.html (3)

The function setPopup was removed. Errors are handled by clipboardError
function, the validation of selection obj happens as soon as it is
retrieved from get_selection.js (2).

If no errors happen, it goes on to convert the selection to an URI in
the async function getFileFromClip

1. content_scripts/get_selection.js
2. popup.js @ storeClipboard
3. popup.html